moved VolumeControl to its own source file
