+++ lg_volume_control/CMakeLists.txt
@@ -5,7 +5,7 @@ find_package(catkin REQUIRED COMPONENTS
   rospy
 )
 
-#catkin_python_setup()
+catkin_python_setup()
 
 catkin_package()
