+++ lg_volume_control/src/lg_volume_control/__init__.py
@@ -1,0 +1,1 @@
+from volume_control import VolumeControl
